sudoku is working

+++ src/parser_camera.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   parser_camera.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jmykkane <jmykkane@student.42.fr>          +#+  +:+       +#+        */
+/*   By: djames <djames@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/22 13:15:41 by joonasmykka       #+#    #+#             */
-/*   Updated: 2023/09/11 12:05:06 by jmykkane         ###   ########.fr       */
+/*   Updated: 2023/09/13 15:38:06 by djames           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 #include "parser.h"
 
-static int	assign_fov(int *fov, char *str)
+static int	assign_fov(float *fov, char *str)
 {
 	int		value;
